pyproject.toml
[project]
name = "gym-tracker"
version = "0.1.0"
description = "A web-based gym progress tracker with AI insights"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.128.0",
    "pydantic>=2.12.5",
    "uvicorn>=0.40.0",
    "firebase-admin>=6.5.0",
    "python-dotenv>=1.0.0",
]
